requirements.txt
Django>=4.2
requests
Pillow
python-dotenv